Bugfix: fix segfault when -t first or -t last used with -iu or -id
closestBed does some extra filtering of the hits returned by BedFile::allHits, but
it was previously assumed that each hit from allHits would be counted by closestBed.

This fix adds a check for the -t to make sure there are some hits in closestB and
updates numOverlaps when hits are filtered.

Maybe it would be better to add orientation-specific search to BedFile::allHits
instead of in closestBed?