Update merge.rst
For sort, should chr9 be before chr10?  

I Added a V to the end of the first sort criteria.  V is for version.  In this case V will sort the numeric section of the first field as a number so that chr9 comes before chr10.  If the first field is just for grouping and the order of chr9 vs chr10 does not matter, then this can be left out.