fix description of option -Sm
Option -Sm is described in the summary table, and in the title line of its description subsection, as retrieving matches on the same strand. However, that's what option -sm does. The actual description is correct, -Sm retrieves matches on the opposite strand. I changed the word 'same' to 'opposite' in the summary table and in the header of the description subsection.